// types/user.ts
export interface User {
    id:            number;
    name:          string;
    email:         string;
    phone:         string;
    password:      string;
    point:         number;
    ref_code:      string;
    link_ref_code: null;
    created_on:    Date;
    updated_on:    null;
    deleted:       number;
}